Enforce pointers for borrow and mutability
When a function receives a struct directly, not a pointer to it, it
signals that it is the responsibility of that function to free the
struct's heap allocated fields. Using pointers signals that the function
is only **borrowing** the struct and won't free it's fields.
A `const` pointer means that the contents behind the reference cannot be
changed. Therefore, when `const` is not used, we should expect the
function to change the contents behind the reference.