Reduce coupling of functions.
Previously the functions `read_..._register()` would have to assume that
the field `removido` had already been read which means the function is
tightly coupled.
There is still the issue that we cannot know if EOF was read in the
first attempt of `fread()` or at the middle of the register, which would
indicate an error. Optimally there would be three possible return values:
ok, read error and eof error.