Purge `common.h` coupling from the csv library.
As described in the description of the csv module, this library SHOULD
NOT have any knowledge of any other non-standard libraries. It is a
fully generic module prepared to work on pretty much ANY environment. If
the inner workings of the module doesn't meet requirements, simply change
the calling functions, and not the module's code! The only allowed
exception is the utils module, which is useful in other contexts as well
and only provides a few simple functions.